# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Construction Coordination Agent",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/profile": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Profile",
        "operationId": "profile_api_profile_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/capabilities": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Health Capabilities",
        "operationId": "health_capabilities_api_capabilities_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "probe",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Probe"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapabilityResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/bim/elements": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Bim",
        "operationId": "bim_api_projects__project_id__bim_elements_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "element_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Element Id"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "name": "location",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Location"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BIMElement"
                  },
                  "title": "Response Bim Api Projects  Project Id  Bim Elements Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/geo": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Geo",
        "operationId": "geo_api_projects__project_id__geo_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/documents": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Documents",
        "operationId": "documents_api_projects__project_id__documents_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentMetadata"
                  },
                  "title": "Response Documents Api Projects  Project Id  Documents Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Import Document",
        "operationId": "import_document_api_projects__project_id__documents_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_import_document_api_projects__project_id__documents_post"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentRun"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}/chunks": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Chunks",
        "operationId": "chunks_api_documents__document_id__chunks_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentChunk"
                  },
                  "title": "Response Chunks Api Documents  Document Id  Chunks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{document_id}/content": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Download Document",
        "operationId": "download_document_api_documents__document_id__content_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "document_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Document Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/projects/{project_id}/search": {
      "get": {
        "tags": [
          "resources"
        ],
        "summary": "Search",
        "operationId": "search_api_projects__project_id__search_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Project Id"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 200,
              "title": "Q"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentChunk"
                  },
                  "title": "Response Search Api Projects  Project Id  Search Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/desktop/shutdown": {
      "post": {
        "tags": [
          "resources"
        ],
        "summary": "Desktop Shutdown",
        "operationId": "desktop_shutdown_api_desktop_shutdown_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentRun": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "event_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Event Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "QUEUED",
              "RUNNING",
              "WAITING_APPROVAL",
              "COMPLETED",
              "FAILED",
              "CANCELLED",
              "EXPIRED"
            ],
            "title": "Status",
            "default": "QUEUED"
          },
          "runtime": {
            "type": "string",
            "title": "Runtime",
            "default": "dbos"
          },
          "runtime_execution_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Runtime Execution Id"
          },
          "generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Generation",
            "default": 0
          },
          "runtime_generation": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Runtime Generation",
            "default": 0
          },
          "category": {
            "type": "string",
            "enum": [
              "coordination",
              "document_parse",
              "bim_import",
              "optimization",
              "vision",
              "embedding_index"
            ],
            "title": "Category",
            "default": "coordination"
          },
          "analysis_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Id"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "event_id",
          "status",
          "runtime",
          "runtime_execution_id",
          "generation",
          "runtime_generation",
          "category",
          "analysis_id",
          "error",
          "created_at",
          "updated_at"
        ],
        "title": "AgentRun"
      },
...
      "BIMElement": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "storey": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storey"
          },
          "space": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Space"
          },
          "properties": {
            "additionalProperties": true,
            "type": "object",
            "title": "Properties"
          },
          "related_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Related Ids",
            "default": []
          },
          "revision": {
            "type": "string",
            "title": "Revision"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "name",
          "type",
          "storey",
          "space",
          "properties",
          "related_ids",
          "revision"
        ],
        "title": "BIMElement"
      },
...
      "Body_import_document_api_projects__project_id__documents_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_import_document_api_projects__project_id__documents_post"
      },
...
      "Capability": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "implementation": {
            "type": "string",
            "title": "Implementation"
          },
          "status": {
            "type": "string",
            "enum": [
              "enabled",
              "available_disabled",
              "unavailable_dependency",
              "unavailable_credential",
              "unhealthy"
            ],
            "title": "Status"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "dependency_available": {
            "type": "boolean",
            "title": "Dependency Available"
          },
          "credential_present": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Credential Present"
          },
          "service_reachable": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Reachable"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "implementation",
          "status",
          "enabled",
          "dependency_available",
          "credential_present",
          "service_reachable",
          "reason",
          "version"
        ],
        "title": "Capability"
      },
...
      "CapabilityResponse": {
        "properties": {
          "capabilities": {
            "items": {
              "$ref": "#/components/schemas/Capability"
            },
            "type": "array",
            "title": "Capabilities"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "capabilities"
        ],
        "title": "CapabilityResponse"
      },
...
      "DocumentChunk": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "text": {
            "type": "string",
            "title": "Text"
          },
          "page": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Page"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "source_hash": {
            "type": "string",
            "title": "Source Hash"
          },
          "parser": {
            "type": "string",
            "title": "Parser"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "text",
          "page",
          "location",
          "source_hash",
          "parser"
        ],
        "title": "DocumentChunk"
      },
...
      "DocumentMetadata": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_hash": {
            "type": "string",
            "title": "Content Hash"
          },
          "parser": {
            "type": "string",
            "title": "Parser"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "project_id",
          "filename",
          "content_hash",
          "parser",
          "created_at"
        ],
        "title": "DocumentMetadata"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ProfileResponse": {
        "properties": {
          "profile": {
            "type": "string",
            "title": "Profile"
          },
          "runtime": {
            "type": "string",
            "title": "Runtime"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning"
          },
          "storage": {
            "type": "string",
            "title": "Storage"
          },
          "database": {
            "type": "string",
            "title": "Database"
          },
          "simulation": {
            "type": "boolean",
            "title": "Simulation"
          },
          "authentication": {
            "type": "string",
            "title": "Authentication"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "profile",
          "runtime",
          "reasoning",
          "storage",
          "database",
          "simulation",
          "authentication"
        ],
        "title": "ProfileResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}